Move actions to its own store. Fix readonly reload bug
