+++ src/lib/editor/stores.ts
@@ -1,10 +1,8 @@
-import { derived, writable, get } from "svelte/store";
+import { derived, writable } from "svelte/store";
 import * as GOBL from "@invopop/gobl-worker";
 import type { GOBLError } from "@invopop/gobl-worker";
 import type * as monaco from "monaco-editor";
 import type { Notification } from "$lib/notifications/index.js";
-import { encodeUTF8ToBase64 } from "$lib/encodeUTF8ToBase64.js";
-import { Severity, createNotification } from "$lib/notifications/index.js";
 
 function createKeypairStore() {
   const { subscribe, set } = writable<GOBL.Keypair | null>(null);
@@ -125,119 +123,4 @@ export function envelopeDocumentSchema(envelope: Envelope | null): string {
     return "";
   }
   return envelope.doc.$schema;
-}
-
-export async function handleBuild() {
-  if (!get(validEditor)) {
-    return;
-  }
-
-  try {
-    const sendData = getGOBLPayload();
-
-    const payload: GOBL.BuildPayload = {
-      data: encodeUTF8ToBase64(sendData),
-      draft: true,
-      envelop: true,
-    };
-    const rawResult = await GOBL.build({ payload });
-    const result = JSON.parse(rawResult);
-
-    envelope.set(result);
-    goblError.set(null);
-
-    createNotification({
-      severity: Severity.Success,
-      message: "Document successfully built.",
-    });
-
-    return { result };
-  } catch (e) {
-    const goblErr = GOBL.parseGOBLError(e);
-    goblError.set(goblErr);
-
-    return {
-      error: goblErr,
-    };
-  }
-}
-
-export async function handleSign() {
-  const keypairValue = get(keypair);
-
-  if (!get(validEditor) || !keypairValue) {
-    return;
-  }
-
-  try {
-    const sendData = getGOBLPayload();
-
-    const payload: GOBL.SignPayload = {
-      data: encodeUTF8ToBase64(sendData),
-      privatekey: keypairValue.private,
-    };
-    const rawResult = await GOBL.sign({ payload });
-    const result = JSON.parse(rawResult);
-
-    envelope.set(result);
-    goblError.set(null);
-
-    createNotification({
-      severity: Severity.Success,
-      message: "Document successfully signed.",
-    });
-
-    return { result };
-  } catch (e) {
-    const goblErr = GOBL.parseGOBLError(e);
-    goblError.set(goblErr);
-
-    return {
-      error: goblErr,
-    };
-  }
-}
-
-export async function handleValidate() {
-  if (!get(validEditor) || !get(envelopeIsSigned)) {
-    return;
-  }
-
-  try {
-    const sendData = getGOBLPayload();
-
-    const payload: GOBL.ValidatePayload = {
-      data: encodeUTF8ToBase64(sendData),
-    };
-    await GOBL.validate({ payload });
-
-    goblError.set(null);
-
-    createNotification({
-      severity: Severity.Success,
-      message: "Document successfully validated.",
-    });
-
-    return { isValid: true };
-  } catch (e) {
-    const goblErr = GOBL.parseGOBLError(e);
-    goblError.set(goblErr);
-
-    return {
-      isValid: false,
-      error: goblErr,
-    };
-  }
-}
-
-function getGOBLPayload() {
-  const editorValue = get(editor);
-  const envelopeValue = get(envelope);
-  const doc = JSON.parse(editorValue || "");
-  if (doc.$schema == envelopeGOBLSchema) {
-    return editorValue || ""; // send as-is
-  }
-  envelopeValue.doc = doc;
-
-  return JSON.stringify(envelopeValue);
 }
